Crate clap_derive

clap_derive

Macro implementation for clap’s derives.

docs.rs

Derive Tutorial

Derive Reference

License

Licensed under either of

Apache License, Version 2.0 (LICENSE-APACHE or https://www.apache.org/licenses/LICENSE-2.0)

MIT license (LICENSE-MIT or https://opensource.org/licenses/MIT)

at your option.

Contribution

Unless you explicitly state otherwise, any contribution intentionally submitted for inclusion in the work by you, as defined in the Apache-2.0 license, shall be dual licensed as above, without any additional terms or conditions.

See CONTRIBUTING for more details.

Derive Macros

ArgEnum

Generates the ValueEnum impl.

Args

Generates the Args impl.

Parser

Generates the Parser implementation.

Subcommand

Generates the Subcommand impl.

ValueEnum

Generates the ValueEnum impl.